fix endgame crash
apparently if the game starts with the last enemy, and player manages to
kill mid-turn, the game doesn't end in a win sometimes, because of the
enemyDead logic. I am guessing this as the reason, because the game was just
crashes with a bad_alloc.

unclear if this changes actual game mechanics.